requirements.txt
hacking>=0.10.0,<0.11
dateutils
